requirements.txt
streamlit==0.65.2
numpy==1.19.1
tensorflow==2.3.1
opencv-python==4.4.0.44
Pillow==7.2.0



